requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
distro==1.8.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
openai==1.3.6
packaging==23.2
pdf2image==1.16.3
Pillow==10.1.0
pydantic==2.5.2
pydantic_core==2.14.5
pytesseract==0.3.10
python-dotenv==1.0.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
pdfplumber==0.10.3
